fix(orders): replace payment method block with new ResourcePaymentMethod component
